requirements.txt
# Dependências do Projeto
pytest-cov
ruff

# CLI do Projeto
.